+++ decidim-proposals/spec/system/search_proposals_spec.rb
@@ -8,5 +8,9 @@ describe "Search proposals", type: :system do
   let!(:searchables) { create_list(:proposal, 3, component: component) }
   let!(:term) { searchables.first.title.split(" ").sample }
 
+  before do
+    searchables.each { |s| s.update(published_at: DateTime.current) }
+  end
+
   include_examples "searchable results"
 end